Use generics for dataset table
