+++ frontend/src/app/components/DatasetCard/DatasetTable.tsx
@@ -25,69 +25,7 @@ import {
   TableRow,
 } from "@/components/ui/table"
 
-
-const keys = ['id', 'amount', 'status', 'email'] as const;
-type ColumnType = Record<typeof keys[number], string>;
-
-const data: ColumnType[] = [
-  {
-    id: "m5gr84i9",
-    amount: "316",
-    status: "success",
-    email: "ken99@yahoo.com",
-  },
-  {
-    id: "3u1reuv4",
-    amount: "242",
-    status: "success",
-    email: "Abe45@gmail.com",
-  },
-  {
-    id: "derv1ws0",
-    amount: "837",
-    status: "processing",
-    email: "Monserrat44@gmail.com",
-  },
-  {
-    id: "5kma53ae",
-    amount: "874",
-    status: "success",
-    email: "Silas22@gmail.com",
-  },
-  {
-    id: "bhqecj4p",
-    amount: "721",
-    status: "failed",
-    email: "carmella@hotmail.com",
-  },
-]
-
-const columns: ColumnDef<ColumnType>[] = keys.map((key) => {
-  return {
-    accessorKey: key,
-    header: ({ column }) => {
-      return (
-        <button
-          className="inline-flex hover:text-gray-800 transition-colors duration-100"
-          onClick={() => column.toggleSorting(column.getIsSorted() === "asc")}
-        >
-          {key}
-          {!column.getIsSorted()
-            ? <ArrowUpDown className="ml-1 size-4" />
-            : column.getIsSorted() === 'asc'
-              ? <ArrowUp className="ml-1 size-4" />
-              : <ArrowDown className="ml-1 size-4" />
-          }
-        </button>
-      )
-    },
-    cell: ({ row }) => (
-      <div>{row.getValue(key)}</div>
-    ),
-  }
-})
-
-export function DataTableDemo() {
+export function DataTableDemo<T>({ data, columns }: { data: T[], columns: ColumnDef<T>[] }) {
   const [sorting, setSorting] = React.useState<SortingState>([])
   const [columnFilters, setColumnFilters] = React.useState<ColumnFiltersState>(
     []
